refactor: Schedule 컨텍스트 타입 분리 및 index파일 생성
