+++ airbnb/src/Components/Schedule/Schedule.tsx
@@ -1,6 +1,6 @@
 import ArrowForwardIosRoundedIcon from '@mui/icons-material/ArrowForwardIosRounded';
 import ArrowBackIosRoundedIcon from '@mui/icons-material/ArrowBackIosRounded';
-import { ScheduleProps, ScheduleProvider } from '@/Contexts/Schedule/context';
+import { ScheduleProps, ScheduleProvider } from '@/Contexts/Schedule';
 import { useCurrentDate } from '@/Hooks/ScheduleHooks';
 import * as S from './Schedule.style';
 import { Calendar } from './Calendar';
